refactor(apps/sveltekit-example-app): move sign-up under account
Also provide a bit of style for the form
