+++ apps/sveltekit-example-app/src/routes/account/sign-up/+page.server.ts
